Increase initial buffer to accommodate META
According to the project gemini specification, the META part
of the response may be 1024 byte long. Adding a two-byte status
code, a single space after that status code and a CR-LF line
terminator at the end, we arrive at up to 1029 bytes that need
to be read to accommodate any valid server response.