Try to extract author and date from url if they are missing
This includes a change to treat empty '%!GEMDOC key' directives
as setting 'key' to an empty value. This will work both if the
equals sign in the '%!GEMDOC' directive is followed only by
whitespace and if the equals sign is missing entirely. Allowing
these empty directives gives users a convenient way to override
the automated date and author extraction from urls.