Move import into main loop to improve performance
Importing HTML and CSS from weasyprint is a rather expensive
operation. Since these two classes are only really needed if
gemdoc is supposed to actually do the gemini to pdf conversion.
This commit moves the actual import down into the main part of
the script, so that it may be skipped if any non-conversion
functionality is requested via command line flags. This way,
these modules don't need to be loaded for simply printing the
default css style or the help message.