Set pdf file identifiers based on content hashes
The ID field in the pdf's trailer dictionary is mandatory for
PDF/A-3B compliance. This commit sets the primary and secondary
file identifier based on the hexadecimal sha256 digest of the
source and rendered representation of the document respectively.
In addition to generating unique ids, this has the additional
benefit that documents that share the same text/gemini source
will also share the same primary file identifier.