Add comments about a85encode breaking images
Running the a85encode filter on images apparently breaks them. In
order to make images work, simply remove the two lines marked with
comments in this commit. Note that the a85encode filter is required
to make the resulting PDF valid utf-8. If it is not run, multiple of
gemdoc's features, such as in-place conversion of polyglot files,
will break.

Once I figure out why the a85encode filter breaks images, I might
be able to come up with a solution that allows for both images and
valid utf-8 polyglot files.